did some stuff. arrow keys
git-svn-id: svn+ssh://archlunatic.dyndns.org/home/svn/repositories/space-engine@15 5191e29d-c056-4a32-ad8b-6a9f2c0e92a1

+++ sprite.cpp
@@ -39,7 +39,7 @@ void sprite::setPivot(unsigned int x, unsigned int y)
     this->pivoty = y;
 }
 
-void sprite::setPosition(int x, int y)
+void sprite::setPosition(double x, double y)
 {
     this->xposition = x/2;//merp? UGLY
     this->yposition = y/2;
